feat: improve telegram logging with retries
